requirements.txt
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=0.24.0
pyserial>=3.5
pandas>=1.3.0